fix: Resolve 181 bugs across entire application
Backend (31 files):
- Add missing db.commit() in approvals, chat, contacts, bim endpoints
- Fix inspection schemas: BaseModel → CamelCaseModel for proper camelCase responses
- Add RBAC permission checks to project delete/update, contact groups, checklists, inspections
- Add security fixes: invalidate old reset tokens, admin self-modification protection
- Replace deprecated datetime.utcnow() with datetime.now(timezone.utc) across 12 files
- Add status guards for equipment/materials submission (draft only)
- Fix RFI: restrict editing sent RFIs, add is_internal to response schema, add validators
- Fix meetings: duplicate attendee check, consistent transaction handling
- Fix BIM: class-level token cache, file size limit, remove storage_path from response
- Fix quantity extraction: null response check, robust JSON parsing, regex markdown strip
- Fix analytics: N+1 query refactored to GROUP BY, budget COUNT vs SUM
- Fix calendar: remove UTC Z suffix from naive datetimes
- Fix tasks: exclude completed from overdue count

Frontend (63 files):
- Fix auth: loading states, double-submit protection, biometric state separation
- Fix RTL: borderInlineEnd, margin-start, insetInlineEnd for Hebrew layout
- Fix analytics: correct RFI data mapping (was showing material data)
- Fix i18n: 100+ new Hebrew/English translation keys, remove hardcoded strings
- Fix validation: camelCase error keys matching form fields
- Fix API: correct endpoint URLs, snake_case field names for backend
- Fix pages: responsive padding, proper error handling, date-based filtering
- Fix MobileChecklistPage: fetch template for subsections, full i18n
- Fix components: ForgeViewer ref stability, IFCViewer cleanup, dead code removal
- Fix contexts: persist project selection, clear cache on logout, auth event system

+++ backend/app/schemas/budget.py
@@ -134,7 +134,7 @@ class BudgetSummaryResponse(CamelCaseModel):
     total_budgeted: Decimal
     total_actual: Decimal
     total_variance: Decimal
-    total_change_orders: Decimal
+    total_change_orders: int
     approved_change_orders: Decimal
     line_item_count: int
     cost_entry_count: int
